@@ -24,10 +24,10 @@
     "jsoninput": "github:cigit/JSONInput",
     "lodash-es": "^4.17.11",
     "prop-types": "^15.7.2",
-    "react": "^16.8.2",
+    "react": "^16.8.3",
     "react-ace": "^6.4.0",
     "react-color": "^2.17.0",
-    "react-dom": "^16.8.2",
+    "react-dom": "^16.8.3",
     "react-sortable-hoc": "^0.8.3",
     "react-widgets": "^4.4.9",
     "recast": "^0.17.3"
@@ -41,7 +41,7 @@
     "@babel/preset-react": "^7.0.0",
     "@types/lodash-es": "^4.17.2",
     "@types/prop-types": "^15.5.9",
-    "@types/react": "^16.8.3",
+    "@types/react": "^16.8.4",
     "@types/react-color": "^2.14.1",
     "@types/react-dom": "^16.8.2",
     "@types/react-sortable-hoc": "^0.6.5",
@@ -49,7 +49,6 @@
     "babel-eslint": "^8.2.3",
     "babel-loader": "^8.0.5",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "copy-webpack-plugin": "^4.6.0",
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.0",
     "eslint": "^4.19.1",
@@ -70,7 +69,7 @@
     "ts-loader": "^5.3.3",
     "tslint": "^5.11.0",
     "tslint-react": "^3.6.0",
-    "typescript": "^3.1.6",
+    "typescript": "^3.3.3333",
     "url-loader": "^1.1.1",
     "webpack": "^4.29.5",
     "webpack-cli": "^3.2.3"
