@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import deb from 'lodash-es/debounce';
 import labeled from '../HOC/labeled';
 import commonView from '../HOC/commonView';
@@ -29,8 +29,8 @@ export const inputStyle = css({
     },
 });
 
-function fromNotToEmpty(value?: void | string | number) {
-    if (value === null || value === undefined) {
+function fromNotToEmpty(value?: string | number) {
+    if (value == null) {
         return '';
     }
     return value;
@@ -49,76 +49,53 @@ interface IStringProps {
     };
     onChange: (value: string) => void;
 }
-class StringView extends React.Component<
-    IStringProps,
-    { value: string | number }
-> {
-    debouncedOnChange: ((value: string) => void) & _.Cancelable;
-    constructor(props: IStringProps) {
-        super(props);
-        this.state = { value: fromNotToEmpty(props.value) };
-        this.handleChange = this.handleChange.bind(this);
-        this.debouncedOnChange = deb(props.onChange, 300);
-    }
-    componentWillReceiveProps(nextProps: IStringProps) {
-        this.setState({ value: fromNotToEmpty(nextProps.value) });
-        if (nextProps.onChange !== this.props.onChange) {
-            this.debouncedOnChange.flush();
-            this.debouncedOnChange = deb(nextProps.onChange, 300);
+
+function StringView(props: IStringProps) {
+    const [value, setValue] = React.useState(fromNotToEmpty(props.value));
+    const debounced = React.useCallback(deb(props.onChange, 300), [
+        props.onChange,
+    ]);
+    React.useEffect(() => () => debounced.flush(), [debounced]);
+
+    function handleChange(event: { target: { value: string } }) {
+        if (event.target.value !== value) {
+            setValue(event.target.value);
+            debounced(event.target.value);
         }
     }
-    componentWillUnmout() {
-        this.debouncedOnChange.flush();
-    }
-    handleChange(
-        event:
-            | React.ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
-            | React.FocusEvent<HTMLTextAreaElement | HTMLInputElement>
-    ) {
-        const value = event.currentTarget.value;
-        const eventType = event.type;
-        this.setState({ value }, () => {
-            if (eventType === 'change') {
-                if (!this.props.blurOnly) {
-                    this.debouncedOnChange(value);
-                }
-            } else {
-                this.props.onChange(value);
-            }
-        });
-    }
-    render() {
-        if (typeof this.props.view.rows === 'number') {
-            return (
-                <textarea
-                    id={this.props.id}
-                    className={inputStyle.toString()}
-                    rows={this.props.view.rows}
-                    onChange={this.handleChange}
-                    onBlur={this.handleChange}
-                    placeholder={this.props.view.placeholder}
-                    value={this.state.value}
-                    disabled={this.props.view.disabled}
-                    readOnly={this.props.view.readOnly}
-                    autoComplete="off"
-                />
-            );
-        }
+
+    React.useEffect(() => setValue(fromNotToEmpty(props.value)), [
+        props.value,
+    ]);
+    if (typeof props.view.rows === 'number') {
         return (
-            <input
-                id={this.props.id}
+            <textarea
+                id={props.id}
                 className={inputStyle.toString()}
-                type="text"
-                placeholder={this.props.view.placeholder}
-                value={this.state.value}
-                onChange={this.handleChange}
-                onBlur={this.handleChange}
-                disabled={this.props.view.disabled}
-                readOnly={this.props.view.readOnly}
+                rows={props.view.rows}
+                onChange={handleChange}
+                onBlur={handleChange}
+                placeholder={props.view.placeholder}
+                value={value}
+                disabled={props.view.disabled}
+                readOnly={props.view.readOnly}
                 autoComplete="off"
             />
         );
     }
+    return (
+        <input
+            id={props.id}
+            className={inputStyle.toString()}
+            type="text"
+            placeholder={props.view.placeholder}
+            value={value}
+            onChange={handleChange}
+            onBlur={handleChange}
+            disabled={props.view.disabled}
+            readOnly={props.view.readOnly}
+            autoComplete="off"
+        />
+    );
 }
-
 export default commonView(labeled(StringView));
